hw1 modified

+++ downloadFile/src/urlThread.java
@@ -1,7 +1,6 @@
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.MalformedURLException;
 import java.net.URL;
 import java.nio.file.Files;
 
@@ -15,12 +14,19 @@ public class urlThread extends Thread {
     @Override
     public void run(){
         try {
-            String extension = strUrl.substring(strUrl.lastIndexOf('/')+1);
-            URL url = new URL(strUrl);
-            InputStream inputStream = url.openStream();
-            Files.copy(inputStream, new File("c:/" + extension).toPath());
+                String extension = strUrl.substring(strUrl.lastIndexOf('/')+1);
+                File file = new File("c:/" + extension);
+                    if(file.exists())
+                        System.out.println("File " + extension + " already exists");
+                    else {
+                        URL url = new URL(strUrl);
+                        InputStream inputStream = url.openStream();
+                        Files.copy(inputStream, new File("c:/" + extension).toPath());
+
+                        System.out.println("File " + extension + " successfully downloaded");
+                    }
         } catch (IOException e) {
-            System.out.println("URL is invalid");
+            System.out.println("URL is invalid " + strUrl);
         }
     }
 }
